(PC-36074): separate top part into a component
