+++ src/features/bookings/components/Ticket/TicketTopPart.native.test.tsx
@@ -1,52 +1,53 @@
 import React from 'react'
 
-import { TicketDisplay } from 'features/bookings/components/Ticket/TicketDisplay'
-import { bookingsSnap } from 'features/bookings/fixtures/bookingsSnap'
+import { TicketTopPart } from 'features/bookings/components/Ticket/TicketTopPart'
 import { subcategoriesMappingSnap } from 'libs/subcategories/fixtures/mappings'
 import { SubcategoriesMapping } from 'libs/subcategories/types'
+import { mockBuilder } from 'tests/mockBuilder'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { render, screen } from 'tests/utils'
-import { Typo } from 'ui/theme'
 
-describe('TicketCutoutContent', () => {
+const offer = mockBuilder.bookingOfferResponse()
+
+describe('TicketTopPart', () => {
   it('should display duo block when offer is duo', async () => {
-    renderTicketDisplay({ isDuo: true })
+    renderTicketTopPart({ isDuo: true })
 
     expect(screen.getByText('Pour deux personnes')).toBeOnTheScreen()
   })
 
   it('should not display duo block when offer is not duo', async () => {
-    renderTicketDisplay({})
+    renderTicketTopPart({})
 
     expect(screen.queryByText('Pour deux personnes')).not.toBeOnTheScreen()
   })
 
   it('should display day block when offer has a day', async () => {
-    renderTicketDisplay({ day: '20 mars 2025' })
+    renderTicketTopPart({ day: '20 mars 2025' })
 
     expect(screen.getByText('20 mars 2025')).toBeOnTheScreen()
   })
 
   it('should not display day block when offer has no day', async () => {
-    renderTicketDisplay({})
+    renderTicketTopPart({})
 
     expect(screen.queryByText('20 mars 2025')).not.toBeOnTheScreen()
   })
 
   it('should display hour block when offer has an hour', async () => {
-    renderTicketDisplay({ hour: '18h56' })
+    renderTicketTopPart({ hour: '18h56' })
 
     expect(screen.getByText('18h56')).toBeOnTheScreen()
   })
 
   it('should not display hour block when offer has no hour', async () => {
-    renderTicketDisplay({})
+    renderTicketTopPart({})
 
     expect(screen.queryByText('18h56')).not.toBeOnTheScreen()
   })
 })
 
-const renderTicketDisplay = ({
+const renderTicketTopPart = ({
   hour,
   day,
   isDuo,
@@ -57,15 +58,14 @@ const renderTicketDisplay = ({
 }) => {
   return render(
     reactQueryProviderHOC(
-      <TicketDisplay
-        title="Super spectacle"
-        hour={hour}
-        day={day}
+      <TicketTopPart
         isDuo={isDuo}
-        offer={bookingsSnap.ongoing_bookings[0].stock.offer}
-        mapping={subcategoriesMappingSnap as SubcategoriesMapping}>
-        <Typo.Body>Partie basse du ticket</Typo.Body>
-      </TicketDisplay>
+        day={day}
+        hour={hour}
+        title="Toto au cirque"
+        offer={offer}
+        mapping={subcategoriesMappingSnap as SubcategoriesMapping}
+      />
     )
   )
 }
